refactor : section 일급컬렉션 구현
